refactor(ai): streamline write tool handling and improve utility functions
- Introduced a centralized module for AI write tools, consolidating related functions and constants.
- Replaced deprecated utility functions with new implementations for checking successful write tool parts.
- Updated the chat and message components to utilize the new utility functions, enhancing code clarity and maintainability.
- Removed obsolete code and tests related to previous write tool implementations, ensuring a cleaner codebase.

This refactor enhances the overall structure of the AI write tool logic and improves the reliability of write operations in the chat interface.
